gpu-worker/vllm/requirements.txt
# vLLM Worker Dependencies
# Large Language Model inference with vLLM

self-hosted>=1.7.0
vllm>=0.6.0
transformers>=4.40.0
accelerate>=0.30.0
